Moved about us to components folder
