+++ src/app/components/about-us/about-us.component.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/custom-bootstrap-variables.scss";
+@import "../../../assets/scss/custom-bootstrap-variables.scss";
 
 .main-div {
     .image-div {
